refactor: handle bg change
